APP-3135 Applied review comments
