+++ symphony-bdk-examples/bdk-core-examples/src/main/java/com/symphony/bdk/examples/ComplexMessageExample.java
@@ -1,6 +1,7 @@
 package com.symphony.bdk.examples;
 
 import static com.symphony.bdk.core.activity.command.SlashCommand.slash;
+import static com.symphony.bdk.core.config.BdkConfigLoader.loadFromSymphonyDir;
 import static java.util.Collections.singletonMap;
 
 import com.symphony.bdk.core.SymphonyBdk;
@@ -9,11 +10,14 @@ import com.symphony.bdk.template.api.Template;
 
 import lombok.extern.slf4j.Slf4j;
 
-@Slf4j
-public class ComplexMessageExample extends BdkExample {
+import java.io.InputStream;
 
-  @Override
-  protected void run(final SymphonyBdk bdk) throws Exception {
+@Slf4j
+public class ComplexMessageExample {
+
+  public static void main(String[] args) throws Exception {
+
+    final SymphonyBdk bdk = new SymphonyBdk(loadFromSymphonyDir("config.yaml"));
 
     final Template template = bdk.messages().templates().newTemplateFromClasspath("/complex-template.ftl");
 
@@ -35,7 +39,7 @@ public class ComplexMessageExample extends BdkExample {
     bdk.datafeed().start();
   }
 
-  public static void main(String[] args) {
-    BdkExample.run(ComplexMessageExample.class);
+  protected static InputStream loadAttachment(String path) {
+    return ComplexMessageExample.class.getResourceAsStream(path);
   }
 }
